+++ tessera-jaxrs/sync-jaxrs/src/main/java/com/quorum/tessera/p2p/PartyStore.java
@@ -1,9 +1,13 @@
-package com.quorum.tessera.discovery;
+package com.quorum.tessera.p2p;
 
 import java.net.URI;
 import java.util.ServiceLoader;
 import java.util.Set;
 
+/*
+ * Support legacy collation of all parties to be added to
+ * party info reponses so nodes learn of nodes.
+ */
 public interface PartyStore {
 
     Set<URI> getParties();
