+++ tessera-jaxrs/sync-jaxrs/src/main/java/com/quorum/tessera/p2p/PartyStoreFactory.java
@@ -1,4 +1,4 @@
-package com.quorum.tessera.discovery;
+package com.quorum.tessera.p2p;
 
 import java.net.URI;
 import java.util.Set;
