+++ tessera-jaxrs/sync-jaxrs/src/main/java/com/quorum/tessera/p2p/SimplePartyStore.java
@@ -1,4 +1,4 @@
-package com.quorum.tessera.discovery;
+package com.quorum.tessera.p2p;
 
 
 import org.slf4j.Logger;
