+++ tessera-jaxrs/sync-jaxrs/src/main/resources/META-INF/services/com.quorum.tessera.p2p.PartyStore
@@ -1,0 +1,1 @@
+com.quorum.tessera.p2p.PartyStoreFactory
