--- tessera-partyinfo/src/main/resources/META-INF/services/com.quorum.tessera.discovery.PartyStore
@@ -1,1 +1,0 @@
-com.quorum.tessera.discovery.PartyStoreFactory
